+++ src/main/java/com/denis/test/server/repository/UserRepository.java
@@ -3,7 +3,6 @@ package com.denis.test.server.repository;
 import com.denis.test.server.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
     UserEntity findUsersByUsernameAndPassword(String username, String password);
 }
